+++ python_base_app/test/test_suite.py
@@ -29,6 +29,7 @@ from python_base_app.test import test_configuration
 from python_base_app.test import test_pytest
 from python_base_app.test import test_base_ci_toolbox
 from python_base_app.test import test_base_web_server
+from python_base_app.test import test_ldap_user_handler
 
 
 def add_test_cases(p_test_suite, p_config_filename=None):
@@ -55,6 +56,10 @@ def add_test_cases(p_test_suite, p_config_filename=None):
     base_test.add_tests_in_test_unit(
         p_test_suite=p_test_suite,
         p_test_unit_class=test_base_web_server.TestBaseWebServer, p_config_filename=p_config_filename)
+
+    base_test.add_tests_in_test_unit(
+        p_test_suite=p_test_suite,
+        p_test_unit_class=test_ldap_user_handler.TestLdapUserHandler, p_config_filename=p_config_filename)
 
 
 def main():